refactor: reorganize substrate parsers into dedicated module structure
Move Zod-based Substrate type parsers from scattered locations into a cohesive
`substrate-parsers/` module with logical organization by functionality:

- **Collections**: Arrays, Sets, BTreeSet, Map parsing (collections.ts)
- **Primitives**: Null, bool, numbers, bigint parsing (primitives.ts)
- **Text/Bytes**: String/Text and Bytes parsing (text.ts, bytes.ts)
- **Advanced Types**: Hash, Address, Option, Struct, Enum parsing
- **Helpers**: ToPrimitive/ToBigInt interfaces for conversion utilities

Key improvements:
- Consolidate scattered parser logic into organized, discoverable modules
- Add comprehensive JSDoc documentation with usage examples
- Introduce new `sb_basic_enum` parser for simple enumeration types
- Centralized re-exports through substrate-parsers/index.ts
- Update all import paths to use new module structure

This refactoring improves code organization and developer experience while
maintaining full API compatibility for existing Substrate type parsing.
