+++ packages/torus-sdk-ts/src/chain/common/types.ts
@@ -1,31 +1,11 @@
 import type { ApiPromise } from "@polkadot/api";
 import type { ApiDecoration } from "@polkadot/api/types";
-import { U8aFixed } from "@polkadot/types";
-import type { H256 } from "@polkadot/types/interfaces";
 import { z } from "zod";
-import { sb_bigint, sb_number } from "../../types/zod.js";
+import { sb_bigint, sb_number } from "../../substrate-parsers/primitives.js";
 
 // ==== API Types ====
 
 export type Api = ApiDecoration<"promise"> | ApiPromise;
-
-// ==== Hash ====
-
-export const sb_h256 = z
-  .custom<H256>((value) => {
-    if (!(value instanceof U8aFixed)) {
-      return false;
-    }
-    if (value.length !== 32) {
-      return false;
-    }
-    return true;
-  })
-  .transform((value) => value.toHex());
-
-export type HexH256 = z.infer<typeof sb_h256>;
-
-export const H256_HEX = z.string().regex(/^0x[a-fA-F0-9]{64}$/);
 
 // ==== Numbers ====
