+++ packages/torus-sdk-ts/src/types/index.ts
@@ -1,4 +1,4 @@
-export * from "./zod.js";
+export * from "../substrate-parsers/index.js";
 export * from "../chain/common/types.js";
 export * from "../chain/common/errors.js";
 
@@ -6,5 +6,3 @@ export * from "./address.js";
 export * from "./namespace/index.js";
 
 export * from "./events.js";
-
-// TODO: organize directory
